requirements/requirements_tmp.txt
alabaster==0.7.12
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.12.13
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
black==22.12.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
crispy-bootstrap5==0.7
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.0.8
django-allauth==0.51.0
django-coverage-plugin==3.0.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-stubs==1.13.1
django-stubs-ext==0.7.0
django_debug_toolbar==3.8.1
docutils==0.19
environ==1.0
exceptiongroup==1.0.4
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.4
filelock==3.8.2
firebase-admin==5.4.0
flake8==6.0.0
flake8-isort==5.0.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
hiredis==2.0.0
httplib2==0.21.0
identify==2.5.10
idna==3.4
imagesize==1.4.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.7.0
isort==5.11.2
jedi==0.18.2
Jinja2==3.1.2
jws==0.1.3
lazy-object-proxy==1.8.0
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
msgpack==1.0.4
mypy==0.982
mypy-extensions==0.4.3
nodeenv==1.7.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==22.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==4.21.12
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pyflakes==3.0.1
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
Pyrebase4==4.6.0
pytest==7.2.0
pytest-django==4.5.2
pytest-sugar==0.9.6
python-dateutil==2.8.2
python-jwt==2.0.1
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stack-data==0.6.2
termcolor==2.1.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.7.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
watchdog==2.2.0
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
